% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_rainbow}
\alias{insert_rainbow}
\title{Insert Rainbow}
\usage{
insert_rainbow(clr)
}
\arguments{
\item{clr}{(character) A vector of one or more colors.}
}
\value{
A character vector of color names.
}
\description{
Take the string "rainbow" and replace it with c("red", "orange", "yellow", "green", "blue", "purple")
}
\examples{

insert_rainbow("rainbow")
insert_rainbow(c("lightsteelblue", "rainbow", "lightsalmon"))
}
