% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_crawl.R
\name{mc_crawl}
\alias{mc_crawl}
\title{Multi-color crawling text (works on single lines of text only)}
\usage{
mc_crawl(txt = "hello world!", colors = NA, dir = -1, pause = 0,
  ...)
}
\arguments{
\item{txt}{(character) Some text to color, stripped of line breaks}

\item{colors}{(character) A vector of colors, defaulting to
the Viridis Plasma palette.

Must all be \href{https://github.com/r-lib/crayon#256-colors}{\code{crayon}}-supported
colors. Any colors in \code{colors()} or hex values (see \code{?rgb})
are fair game.}

\item{dir}{(numeric: -1 or 1) direction of colors, defaults to brighter to darker}

\item{pause}{(numeric) Pause between characters in seconds, default=0}

\item{...}{Further args.}
}
\value{
A string
}
\description{
Multi-color crawling text (works on single lines of text only)
}
\details{
This function requires as many colors as there are characters in your string and
prints them one at a time.

It cannot be used with RGUI (R.app on some systems).
}
\examples{
\dontrun{
mc_crawl()

mc_crawl("It was a dark and stormy night")

options("keep.source=F")
cat("\\014")
mc_crawl('A long time ago in a galaxy far, far away...')
cat('\\n')
cat('\\n')
mc_crawl("It is a period of civil war. Rebel spaceships, striking from a hidden base,")
cat('\\n')
mc_crawl('have won their first victory against the evil Galactic Empire.',dir=1)


}
}
