% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_color.R
\name{multi_color}
\alias{multi_color}
\title{Multi-color text}
\usage{
multi_color(txt = "hello world!", colors = c("red", "orange", "yellow",
  "green", "blue", "purple"), type = "message", ...)
}
\arguments{
\item{txt}{(character) Some text to color.}

\item{colors}{(character) A vector of colors, defaulting to
c("red", "orange", "yellow", "green", "blue", "purple")}

\item{type}{(character) Message (default), warning, or string}

\item{...}{Further args.}
}
\value{
A string if \code{type} is "string", or colored
text if type is "message" or "warning"
}
\description{
Multi-color text
}
\details{
This function evenly (ish) divides up your string into
these colors in the order they appear in \code{colors}.

Any colors in \code{colors()} or hex values (see \code{?rgb})
are fair game.
}
\examples{
multi_color("ahoy")

multi_color(colors = c(rgb(0.1, 0.2, 0.5),
                       "yellow",
                       rgb(0.2, 0.9, 0.1)))

multi_color(
  cowsay::animals[["buffalo"]],
  c("mediumorchid4", "dodgerblue1", "lemonchiffon1"))

multi_color(cowsay:::rms, sample(colors(), 10))
}
