% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{multicolor_logo}
\alias{multicolor_logo}
\title{The multicolor package logo}
\usage{
multicolor_logo(colors = "random", ...)
}
\arguments{
\item{colors}{Vector of colors for the logo. Defaults to "random" which randomly selects one of the \code{palette}s.}

\item{...}{Arguments passed to \code{multi_color}.}
}
\value{
The colorful package logo on load
}
\description{
The multicolor package logo
}
\details{
This function displays the multicolor package logo
in a randomly selected color palette from a pre-selected list of colors.
}
\examples{
multicolor_logo()
multicolor_logo(recycle_chars = TRUE)
multicolor_logo(colors = c("red", "blue"))
}
