% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nix_first_newline}
\alias{nix_first_newline}
\title{Remove the first instance of a newline from a string}
\usage{
nix_first_newline(s)
}
\arguments{
\item{s}{(character) A string}
}
\value{
A string with the first instance of a newline removed.
}
\description{
Remove the first instance of a newline from a string
}
\examples{
nix_first_newline("onetwo\\nthree\\nfour")

# Nothing to remove
nix_first_newline("fivesixseven")
}
