% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{palettes}
\alias{palettes}
\title{Out-of-the-box Color Palettes}
\format{An object of class \code{list} of length 3.}
\usage{
palettes
}
\value{
A character vector of color values.
}
\description{
Take the string "rainbow" and replace it with c("red", "orange", "yellow", "green", "blue", "purple")
}
\examples{

multi_color(things$cat, colors = palettes$lacroix)
}
\keyword{datasets}
