% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newhope.R
\docType{data}
\name{starwars_intro}
\alias{starwars_intro}
\title{Star Wars a New Hope Intro}
\format{An object of class \code{character} of length 1.}
\usage{
starwars_intro
}
\description{
The intro to Episode IV, for use in multicoloring experiments
}
\keyword{datasets}
