% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{crawl}
\alias{crawl}
\title{Multi-color crawling text}
\usage{
crawl(txt = "hello world!", colors = NULL, recycle_chars = FALSE,
  viridis_dir = -1, pause = 0.05, ...)
}
\arguments{
\item{txt}{(character) Some text to color, stripped of line breaks}

\item{colors}{(character) A vector of colors to color each individual character, defaulting to
the Viridis Plasma palette.

Must all be \href{https://github.com/r-lib/crayon#256-colors}{\code{crayon}}-supported
colors. Any colors in \code{colors()} or hex values (see \code{?rgb})
are fair game.}

\item{recycle_chars}{(logical) Should the vector of colors supplied apply to the entire string or
should it apply to each individual character and be recycled?}

\item{viridis_dir}{(numeric: -1 or 1) For Viridis colors, direction of colors to be applied.
Defaults to -1, i.e. brighter to darker.}

\item{pause}{(numeric) Seconds to pause between characters in seconds.}

\item{...}{Further args.}
}
\value{
A string, printed in \code{colors} with \code{pause} seconds between printing each character.
}
\description{
This function works on single lines of text only.
It produces a gif-like representation of the text unfolding from left to right.
}
\details{
This function requires as many colors as there are characters in your string and
prints them one at a time. Provided colors will be recycled if needed.

It cannot be used with RGUI (R.app on some systems) or other environments that do not support
colored text. See \code{multicolor:::use_color} for more info.
}
\examples{
\dontrun{
crawl()

crawl("It was a dark and stormy night")

crawl("Taste the rainbow", colors = "rainbow")

options("keep.source = FALSE")
cat("\\014")
crawl('A long time ago in a galaxy far, far away...')
cat('\\n')
cat('\\n')
crawl("It is a period of civil war. Rebel spaceships, striking from a hidden base,")
cat('\\n')
crawl('have won their first victory against the evil Galactic Empire.', viridis_dir = 1)
}
}
