% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_color.R
\name{multi_color}
\alias{multi_color}
\title{Multi-color text}
\usage{
multi_color(txt = "hello world!", colors = "rainbow",
  type = "message", direction = "vertical", recycle_chars = FALSE,
  ...)
}
\arguments{
\item{txt}{(character) Some text to color.}

\item{colors}{(character) A vector of colors, defaulting to
"rainbow", i.e. c("red", "orange", "yellow", "green", "blue", "purple").

Must all be \href{https://github.com/r-lib/crayon#256-colors}{\code{crayon}}-supported
colors. Any colors in \code{colors()} or hex values (see \code{?rgb})
are fair game.}

\item{type}{(character) Message (default), warning, string or \code{\link{crawl}}.}

\item{direction}{(character) How should the colors be spread? One of
"horizontal" or "vertical".}

\item{recycle_chars}{(logical) Should the vector of colors supplied apply to the entire string or
should it apply to each individual character and be recycled?}

\item{...}{Further args.}
}
\value{
A string if \code{type} is "string", or colored
text if type is "message" or "warning"
}
\description{
Multi-color text
}
\details{
This function evenly (ish) divides up your string into
these colors in the order they appear in \code{colors}.

It cannot be used with RGUI (R.app on some systems).
}
\examples{
\dontrun{
multi_color()

multi_color("ahoy")

multi_color("taste the rainbow",
            c("mediumpurple",
              "rainbow",
             "cyan3"))

multi_color(colors = c(rgb(0.1, 0.2, 0.5),
                       "yellow",
                       rgb(0.2, 0.9, 0.1)))

multi_color(
  cowsay::animals[["buffalo"]],
  c("mediumorchid4", "dodgerblue1", "lemonchiffon1"))

multi_color(cowsay:::rms, sample(colors(), 10))

# Mystery Bulgarian animal
multi_color(things[[sample(length(things), 1)]],
            c("white", "darkgreen", "darkred"),
            direction = "horizontal")

# Mystery Italian animal
multi_color(things[[sample(length(things), 1)]],
            c("darkgreen", "white", "darkred"),
            direction = "vertical")
}
}
