% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_color.R
\name{multi_colour}
\alias{multi_colour}
\title{Multi-colour text}
\usage{
multi_colour(txt = "hello world!", colors = "rainbow",
  type = "message", direction = "vertical", recycle_chars = FALSE,
  add_leading_newline = TRUE, ...)
}
\arguments{
\item{txt}{(character) Some text to colour.}

\item{colors}{(character) A vector of colours, defaulting to
"rainbow", i.e. c("red", "orange", "yellow", "green", "blue", "purple").

Must all be \href{https://github.com/r-lib/crayon#256-colours}{\code{crayon}}-supported
colours. Any colours in \code{colors()} or hex values (see \code{?rgb})
are fair game.}

\item{type}{(character) Message (default), warning, or string.}

\item{direction}{(character) How should the colours be spread? One of
"horizontal" or "vertical".}

\item{recycle_chars}{(logical) Should the vector of colours supplied apply to the entire string or
should it apply to each individual character (if \code{direction} is vertical)
or line (if \code{direction} is horizontal), and be recycled?}

\item{...}{Further args.}
}
\value{
A string if \code{type} is "string", or coloured
text if type is "message" or "warning"
}
\description{
Multi-colour text
}
\details{
This function evenly (ish) divides up your string into
these colours in the order they appear in \code{colours}.

It cannot be used with RGUI (R.app on some systems).
}
\examples{
\dontrun{
multi_colour()

multi_colour("ahoy")

multi_colour("taste the rainbow",
            c("mediumpurple",
              "rainbow",
             "cyan3"))

multi_colour(colours = c(rgb(0.1, 0.2, 0.5),
                       "yellow",
                       rgb(0.2, 0.9, 0.1)))

multi_colour(
  cowsay::animals[["buffalo"]],
  c("mediumorchid4", "dodgerblue1", "lemonchiffon1"))

multi_colour(cowsay:::rms, sample(colours(), 10))

# Mystery Bulgarian animal
multi_colour(things[[sample(length(things), 1)]],
            c("white", "darkgreen", "darkred"),
            direction = "horizontal")

# Mystery Italian animal
multi_colour(things[[sample(length(things), 1)]],
            c("darkgreen", "white", "darkred"),
            direction = "vertical")
}
}
