% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{multicolor_logo}
\alias{multicolor_logo}
\title{The multicolor package logo}
\usage{
multicolor_logo(...)
}
\arguments{
\item{...}{Arguments passed to \code{multi_color}.}
}
\description{
The multicolor package logo
}
\details{
This function displays the multicolor package logo
in a randomly selected color palette from a pre-selected list of colors.
}
\examples{
multicolor_logo()
}
