% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{multicolor_logo}
\alias{multicolor_logo}
\title{The multicolor package logo}
\usage{
multicolor_logo(display_in_knitr = FALSE, ...)
}
\arguments{
\item{...}{Arguments passed to \code{multi_color}.}
}
\description{
The multicolor package logo
}
\examples{
multicolor_logo()
}
