% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/things.R
\docType{data}
\name{things}
\alias{things}
\title{Things}
\format{An object of class \code{list} of length 43.}
\usage{
things
}
\description{
Named vector of animals and other characters e.g. Yoda, from the \code{cowsay} package
}
\details{
\code{things} is a named character list of ASCII animals and characters.
}
\examples{
things[["turkey"]]
things[["chuck"]] \%>\% cat()
cowsay::animals[3] \%>\% cat()
names(things)
multi_color(things[["stretchycat"]])  # To say something, use the cowsay package
}
\keyword{datasets}
