% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textshapes.R
\name{triangle_string}
\alias{triangle_string}
\title{Turn strings into triangle-shaped strings}
\usage{
triangle_string(string, maxlen = 1, step = +1, display = FALSE)
}
\arguments{
\item{string}{(character) Some text to reshape into a triangle.}

\item{maxlen}{(integer) width of top of triangle, defaults to 1, set larger than 1 for downward-pointing triangle}

\item{step}{(integer) Number of characters to expand or contract triangle width per line (set to negative for downward-pointing triangle)}

\item{display}{(logical) Returns string invisibly if FALSE (default), set to TRUE to display returned string}
}
\value{
A string
}
\description{
Turn strings into triangle-shaped strings
}
\details{
Use positive step argument for upward pointing triangle
and negative step and wider maxlen for downward pointing triangle.
}
